prism gate card 복원
